+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-01-?? (TBD)
+### Added
+- New `Lum\App` static class with methods for making
+  common app bootstrap processes simpler.
+### Changed
+- Added support for `lum-mailer` version `3.x` which is a rewrite
+  that breaks compatibility with the older versions.
+
 ## [1.1.0] - 2024-09-19
 ### Fixed
 - A syntax error in `Common/Auth_Token`
@@ -30,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - All DB-specific `Models` are in their own packages now.
 - Auth plugins are in the `lum-auth` package now.
 
-[Unreleased]: https://github.com/supernovus/lum.app.php/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/supernovus/lum.app.php/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/supernovus/lum.app.php/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/supernovus/lum.app.php/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/supernovus/lum.app.php/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/supernovus/lum.app.php/releases/tag/v1.0.0